requirements.txt
gradio
Pillow
numpy
pandas
datasets
scikit-learn
tqdm
pytesseract
transformers
